+++ effectful/internals/disassembler.py
